Generating Random Variables

Fortuna.jl package builds its capacity to generate random variables using generaterv() function by utilizing the widely-adopted Distributions.jl package, enabling seamless integration with other probabilistic programming Julia packages such as Turing.jl and RxInfer.jl. However, unlike Distributions.jl package, Fortuna.jl allows you to generate random variables not only using their parameters, but also using their moments, which often useful in the field of Structural and System Reliability Analysis.

Fortuna.generaterv — Method

generaterv(DistributionName::String, DefineBy::String, Values::Union{Real,Vector{<:Real}})

The function generates random variables given the distribution name and either its moments or parameters.

If DefineBy = "M", moments of the random variable must be provided in Values.

If DefineBy = "P", parameters of the random variable must be provided in Values.

Generating Random Variables Using Moments

# Generate a lognormally distributed random variable R
# with mean (μ) of 15 and standard deviation (σ) of 10:
R = generaterv("Lognormal", "M", [15, 10])
println("μ = $(mean(R))")
println("σ = $(std(R))")

μ = 15.0
σ = 10.0

Generating Random Variables Using Parameters

# Generate a gamma-distributed random variable Q
# with shape parameter (α) of 16 and scale parameter (θ) of 0.625:
Q = generaterv("Gamma", "P", [16, 0.625])
println("α = $(params(Q)[1])")
println("θ = $(params(Q)[2])")

α = 16.0
θ = 0.625

Supported Random Variables

Fortuna.jl package currently supports the following distributions:

Exponential

Gamma

Gumbel

Lognormal

Normal

Poisson

Uniform

Weibull

Tip

If you want to define a random variable that is not supported by Fortuna.jl package, please raise an issue on the Github Issues page.