Second-Order Reliability Methods

The SORM is an improvement over the FORM by accounting for the curved nature of the failure boundary given by $g(\vec{X}) = 0$ around the design point $\vec{x}^{*}$; thus, providing a better approximation of the probability of failure $P_{f}$.

Curve-Fitting Method

Fortuna.jl package implements the CF method that fits a hyper-paraboloid surface with a vertex at the design point $\vec{x}^{*}$ and the principal curvatures matching the principal curvatures of the failure boundary given by $g(\vec{X}) = 0$ at that point. The probabilities $P_{f}$ of failure are estimated using Hohenbichler and Rackwitz (1988) and Breitung (1984) approximations of the exact solution provided by Tvedt (1990). The calculated probabilities of failure $P_{f}$ are then used to estimate the generalized reliability indices $\beta$, which account for the curved nature of the failure boundary given by $g(\vec{X}) = 0$ around the design point $\vec{x}^{*}$.

Point-Fitting Method

Warning

This feature is currently under development.

API

Fortuna.SORM — Type

struct SORM <: AbstractReliabililyAnalysisMethod

Type used to perform reliability analysis using Second-Order Reliability Method (SORM).

Submethod::Fortuna.SORMSubmethod

Fortuna.solve — Method

solve(Problem::ReliabilityProblem, AnalysisMethod::SORM)

Function used to solve reliability analysis using Second-Order Reliability Method (SORM).