Second-Order Reliability Methods

The SORM is an improvement over the FORM by accounting for the curved nature of the failure boundary given by $g(\vec{X}) = 0$ around the design point $\vec{x}^{*}$; thus, providing a better approximation of the probability of failure $P_{f}$.

Curve-Fitting Method

Fortuna.jl package implements the CF method that fits a hyper-paraboloid surface with a vertex at the design point $\vec{x}^{*}$ and the principal curvatures matching the principal curvatures of the failure boundary given by $g(\vec{X}) = 0$ at that point. The probabilities $P_{f}$ of failure are estimated using Hohenbichler and Rackwitz (1988) and Breitung (1984) approximations of the exact solution provided by Tvedt (1990). The calculated probabilities of failure $P_{f}$ are then used to estimate the generalized reliability indices $\beta$, which account for the curved nature of the failure boundary given by $g(\vec{X}) = 0$ around the design point $\vec{x}^{*}$.

Point-Fitting Method

Warning

This feature is currently under development.

API

Fortuna.SORM — Type

struct SORM <: AbstractReliabililyAnalysisMethod

A custom type used by Fortuna.jl to perform reliability analysis using Second-Order Reliability Methods.

Submethod: Analysis method that falls under the category of Second-Order Reliability Methods.

Fortuna.SORMSubmethod — Type

abstract type SORMSubmethod end

A custom abstract supertype used by Fortuna.jl to define various types of Second-Order Reliability Methods.

Fortuna.analyze — Method

analyze(Problem::ReliabilityProblem, AnalysisMethod::SORM)

The function solves the provided reliability problem using any submethod that falls under a broader category of Second-Order Reliability Methods (SORM).